requirements.txt
python-telegram-bot==4.1.1
google-generativeai
bs4
vnstock
telebot
markdown2
pytz
pandas
pillow